+++ diff_diff/__init__.py
@@ -214,7 +214,7 @@ Bacon = BaconDecomposition
 EDiD = EfficientDiD
 ETWFE = WooldridgeDiD
 
-__version__ = "2.9.1"
+__version__ = "3.0.0"
 __all__ = [
     # Estimators
     "DifferenceInDifferences",